Refactored Tracker and Board classes
Boards are now rendered in a slightly less rigid format that allows for toggling board visibility. If a board is not set to visible, it will not take a spot in the grid.
The Tracker is also now passed into board classes directly to be stored as the parent object.